move `Promise.try` to stable ES
