+++ packages/core-js-compat/src/modules-by-versions.mjs
@@ -267,5 +267,6 @@ export default {
     'es.iterator.some',
     'es.iterator.take',
     'es.iterator.to-array',
+    'es.promise.try',
   ],
 };
