+++ packages/core-js/es/promise/index.js
@@ -5,6 +5,7 @@ require('../../modules/es.object.to-string');
 require('../../modules/es.promise');
 require('../../modules/es.promise.all-settled');
 require('../../modules/es.promise.any');
+require('../../modules/es.promise.try');
 require('../../modules/es.promise.with-resolvers');
 require('../../modules/es.promise.finally');
 require('../../modules/es.string.iterator');
