+++ packages/core-js/stable/promise/try.js
@@ -1,0 +1,4 @@
+'use strict';
+var parent = require('../../es/promise/try');
+
+module.exports = parent;
