+++ packages/core-js/stage/3.js
@@ -8,7 +8,6 @@ require('../proposals/explicit-resource-management');
 require('../proposals/float16');
 require('../proposals/json-parse-with-source');
 require('../proposals/math-sum');
-require('../proposals/promise-try');
 require('../proposals/regexp-escaping');
 // TODO: Obsolete versions, remove from `core-js@4`
 require('../proposals/array-grouping-stage-3');
