+++ tests/compat/tests.js
@@ -1022,6 +1022,13 @@ GLOBAL.tests = {
   }],
   'es.promise.reject': PROMISES_SUPPORT,
   'es.promise.resolve': PROMISES_SUPPORT,
+  'es.promise.try': [PROMISES_SUPPORT, function () {
+    var ACCEPT_ARGUMENTS = false;
+    Promise['try'](function (argument) {
+      ACCEPT_ARGUMENTS = argument === 8;
+    }, 8);
+    return ACCEPT_ARGUMENTS;
+  }],
   'es.promise.with-resolvers': [PROMISES_SUPPORT, function () {
     return Promise.withResolvers;
   }],
@@ -1773,13 +1780,6 @@ GLOBAL.tests = {
   'esnext.number.from-string': function () {
     return Number.fromString;
   },
-  'esnext.promise.try': [PROMISES_SUPPORT, function () {
-    var ACCEPT_ARGUMENTS = false;
-    Promise['try'](function (argument) {
-      ACCEPT_ARGUMENTS = argument === 8;
-    }, 8);
-    return ACCEPT_ARGUMENTS;
-  }],
   'esnext.regexp.escape': function () {
     return RegExp.escape('ab') === '\\x61b';
   },
