+++ tests/unit-global/es.promise.try.js
