+++ tests/unit-pure/es.promise.try.js
@@ -1,4 +1,4 @@
-import Promise from 'core-js-pure/full/promise';
+import Promise from 'core-js-pure/es/promise';
 
 QUnit.test('Promise.try', assert => {
   assert.isFunction(Promise.try);
